+++ plugin_ide.core/src/com/github/rustdt/ide/core/engine/RustSourceModelManager.java
@@ -10,16 +10,28 @@
  *******************************************************************************/
 package com.github.rustdt.ide.core.engine;
 
+import java.nio.file.Path;
+
+import org.eclipse.core.resources.IProject;
+
+import com.github.rustdt.ide.core.operations.RustSDKPreferences;
+import com.github.rustdt.tooling.ops.RustParseDescribeParser;
+
+import melnorme.lang.ide.core.LangCore;
 import melnorme.lang.ide.core.engine.SourceModelManager;
-import melnorme.lang.tooling.ElementAttributes;
-import melnorme.lang.tooling.ast.SourceRange;
+import melnorme.lang.ide.core.operations.AbstractToolManager;
+import melnorme.lang.ide.core.utils.ResourceUtils;
 import melnorme.lang.tooling.structure.SourceFileStructure;
-import melnorme.lang.tooling.structure.StructureElement;
-import melnorme.lang.tooling.structure.StructureElementKind;
-import melnorme.utilbox.collections.ArrayList2;
+import melnorme.utilbox.concurrency.OperationCancellation;
+import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.misc.Location;
+import melnorme.utilbox.misc.PathUtil;
+import melnorme.utilbox.misc.StringUtil;
+import melnorme.utilbox.process.ExternalProcessHelper.ExternalProcessResult;
 
 public class RustSourceModelManager extends SourceModelManager {
+	
+	protected final AbstractToolManager toolManager = LangCore.getToolManager();
 	
 	public RustSourceModelManager() {
 	}
@@ -29,11 +41,35 @@ public class RustSourceModelManager extends SourceModelManager {
 		return new StructureUpdateTask(structureInfo) {
 			@Override
 			protected SourceFileStructure createNewData() {
-				SourceRange sr = new SourceRange(0, source.length());
-				StructureElement element = new StructureElement("Outline_Not_Supported", sr, sr, 
-					StructureElementKind.MOD, new ElementAttributes(null), null, null);
-				Location fileLocation = structureInfo.getLocation();
-				return new SourceFileStructure(fileLocation, new ArrayList2<StructureElement>(element), null);
+				
+				ExternalProcessResult describeResult;
+				
+				Location location = structureInfo.getLocation();
+				IProject project = location == null ? null : ResourceUtils.getProject(location);
+				try {
+					/* FIXME: derived value pref */
+					Path path = PathUtil.createPath(RustSDKPreferences.RAINICORN_PATH.getEffectiveValue(project));
+					
+					ProcessBuilder pb = toolManager.createToolProcessBuilder(project, path);
+					describeResult = toolManager.runEngineTool(pb, source, cm);
+				} catch(OperationCancellation e) {
+					return null;
+				} catch(CommonException ce) {
+					toolManager.logAndNotifyError("Error running parse-describe process:", ce.toStatusError());
+					return null;
+				}
+				
+				Location fileLocation = location;
+				
+				String describeOutput = describeResult.getStdOutBytes().toString(StringUtil.UTF8);
+				
+				try {
+					return new RustParseDescribeParser(fileLocation, source).parse(describeOutput);
+				} catch(CommonException ce) {
+					toolManager.logAndNotifyError("Error reading parse-describe output:", ce.toStatusError());
+					return null;
+				}
+				
 			}
 		};
 	}